feat: separate message handlers into a file
